+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `ErasedSet::shrink_to`
+- `ErasedSet::get_or_insert`
 
 ## [0.6.0] - 2022-05-11